+++ babel.config.js
@@ -3,10 +3,6 @@ module.exports = (api) => {
         plugins: [
             ["@babel/plugin-transform-runtime"],
             ['@babel/plugin-proposal-export-namespace-from'],
-            [
-                '@babel/plugin-proposal-class-properties',
-                {loose: true},
-            ]
         ],
         presets: [
             [
@@ -22,10 +18,6 @@ module.exports = (api) => {
         plugins: [
             ["@babel/plugin-transform-runtime"],
             ['@babel/plugin-proposal-export-namespace-from'],
-            [
-                '@babel/plugin-proposal-class-properties',
-                {loose: true},
-            ]
         ],
         presets: [
             [